requirements.txt
python-dotenv>=1.0.0
schedule>=1.2.0
speedtest-cli>=2.1.3
pandas>=2.0.0
matplotlib>=3.7.0
